% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_model.R
\name{anova_lrt}
\alias{anova_lrt}
\title{Produce Table of ANOVA LRT by iterative over the terms}
\usage{
anova_lrt(form_str, data = nhanes)
}
\arguments{
\item{form_str}{string of the formula}

\item{data}{data set}
}
\value{
results as data frame
}
\description{
Produce Table of ANOVA LRT by iterative over the terms
}
\examples{
lm_str <- 'diastolic ~ RIDAGEYR*RIAGENDR + BMXBMI + RIDRETH1'
anova_lrt(lm_str)
}
