% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_model.R
\name{cv_base_m}
\alias{cv_base_m}
\title{Cross Validation for Base Models}
\usage{
cv_base_m(model_list, label, group_col, df = nhanes)
}
\arguments{
\item{model_list}{a list of strings for the models with names}

\item{label}{outcomes}

\item{group_col}{columns for group data in CV}

\item{df}{data set}
}
\value{
MSE matrix
}
\description{
Cross Validation for Base Models
}
\examples{
lm_str <- "diastolic ~ RIDAGEYR*RIAGENDR + BMXBMI + RIDRETH1"
ns_str <- "diastolic ~ ns(RIDAGEYR, knots = seq(30, 80, by = 10), Boundary.knots=c(20,90)) * RIAGENDR + ns(BMXBMI,knots = c(seq(15, 45, by = 5),seq(45,65,by=10)),Boundary.knots=c(10,85)) + RIDRETH1"
model_list <- c(lm_str,ns_str)
names(model_list) <- c("linear","spline")
mse_mtx <- cv_base_m(model_list,label="diastolic",group_col="years",df=nhanes)
}
