% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enwas.R
\name{enwas}
\alias{enwas}
\title{ENWAS main function}
\usage{
enwas(
  i,
  ph.list,
  data,
  model_var,
  gen_cat,
  outcome,
  sex_spec_cat,
  ss.out,
  ss.out.cat,
  n_cutoff
)
}
\arguments{
\item{i}{index}

\item{ph.list}{list of variables to run through as independent variables}

\item{data}{dataframe with ID, outcome,age (age centered at 50),exposures,
poststratification variable}

\item{model_var}{character vector, list of variable to include in model}

\item{gen_cat}{name of post stratification variable}

\item{outcome}{name of outcome variable}

\item{sex_spec_cat}{if sex specific, name of post-stratification variable}

\item{ss.out}{"Yes" or "No", is outcome sex-specifi}

\item{ss.out.cat}{"M/F", "M", "F"}

\item{n_cutoff}{minimum sample size required to run model.}
}
\value{
enwas models
}
\description{
Environment Wide Associate
}
\examples{
> enwas(i, ph.list, data, model_var, gen_cat, outcome, sex_spec_cat,
> ss.out, ss.out.cat, n_cutoff)
}
