% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enwas_plot.R
\name{forest_plot}
\alias{forest_plot}
\title{Forest Plot for single model}
\usage{
forest_plot(xwas_result)
}
\arguments{
\item{xwas_result}{the EnWAS result data frame}
}
\value{
plot result
}
\description{
Forest Plot for single model
}
\examples{
forest_plot(linear_res$enwas_res)
}
