% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enwas_plot.R
\name{forest_plot_mult}
\alias{forest_plot_mult}
\title{Plot Forest for Muliple Models}
\usage{
forest_plot_mult(xwas_result_list)
}
\arguments{
\item{xwas_result_list}{list of the EnWAS result data frames}
}
\value{
plot multiple model results in a single image
}
\description{
Plot Forest for Muliple Models
}
\examples{
forest_plot_mult(list(linear = linear_res$enwas_res,
                               ns = ns_res$enwas_res))
}
