% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enwas.R
\name{invNorm}
\alias{invNorm}
\title{Inverse Normal Transformation}
\usage{
invNorm(x)
}
\arguments{
\item{x}{}
}
\value{
transformed data
}
\description{
Inverse Normal Transformation
}
\examples{
invNorm(nhanes$BMXWAIST)
}
