% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nhanes}
\alias{nhanes}
\title{National Health and Nutrition Examination Survey (NHANES)}
\format{
A data frame with 36790 rows and 61 variables:
\describe{
\item{BMXHT}{Standing Height (cm)}
\item{BMXBMI}{Body Mass Index (kg/m**2)}
....
}
}
\source{
\url{https://www.cdc.gov/nchs/nhanes/index.htm}
}
\usage{
nhanes
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
