% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enwas_plot.R
\name{plot_bins}
\alias{plot_bins}
\title{Plot binned value for single data frame}
\usage{
plot_bins(
  x,
  y,
  data,
  xlab = "Value",
  ylab = "Binned y",
  title = "linear",
  nobsperbin = 600
)
}
\arguments{
\item{x}{x values}

\item{y}{y values}

\item{data}{data frame}

\item{xlab}{x label text}

\item{ylab}{y label text}

\item{title}{plot title}

\item{nobsperbin}{expected number of data points in per bins}
}
\value{
plot results
}
\description{
The y values are binned with respected to x values
}
\examples{
demo = nhanes("DEMO_C")
plot_bins(x=RIDAGEYR,y=INDFMPIR,data = demo)
}
