% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enwas_plot.R
\name{plot_bins}
\alias{plot_bins}
\title{Plot binned value for single data frame}
\usage{
plot_bins(
  x,
  y,
  xlab = "Value",
  ylab = "Binned y",
  title = "linear",
  nbin = 600
)
}
\arguments{
\item{x}{x values}

\item{y}{y values}

\item{xlab}{x label text}

\item{ylab}{y label text}

\item{title}{plot title}

\item{nbin}{expected data points in each bin}
}
\value{
plot results
}
\description{
The y values are binned with respected to x values
}
\examples{
plot_bins(nhanes$BMXBMI, lm_base$residuals,xlab="BMI(kg/m²)", y="Binned residuals", snbin=300)
}
